commented out all synchronized blocks for better debugging (see details below; `synchronized` is not needed with `GatedThread`s anyway)
We use IntelliJ IDEA, and we have a multi-threaded app (designed in such a way that at any moment, only one thread is effectively executing). We run the app in the debugger, and we have an exception breakpoint on any exception. Sometimes, it works: when there's an exception, the debugger stops at the line with the exception, and we get the whole context of it. Sometimes, it doesn't: the debugger stops at the end of a `synchronized (...) { ... }` block, and among variables, there's `Exception`. The `Exception` has `stackTrace`, which helps, but not to the extent of stopping the line of the exception itself.